+++ package.json
@@ -25,7 +25,6 @@
     "@react-navigation/stack": "^6.3.23",
     "@reduxjs/toolkit": "^2.2.1",
     "@tanstack/react-query": "^5.24.1",
-    "@tanstack/react-query-devtools": "^5.24.1",
     "axios": "^1.6.7",
     "iamport-react-native": "^2.0.11",
     "lottie-ios": "4.4.1",
@@ -33,7 +32,7 @@
     "native-base": "^3.4.28",
     "react": "18.2.0",
     "react-error-boundary": "^4.0.12",
-    "react-native": "0.73.4",
+    "react-native": "0.73.5",
     "react-native-collapsible": "^1.6.1",
     "react-native-dropdown-picker": "^5.4.6",
     "react-native-fast-image": "^8.6.3",
